requirements.txt
opencv-contrib-python
Pillow
numpy
matplotlib
customtkinter